requirements.txt
Django==3.2.7
gunicorn==20.1.0
djangorestframework==3.12.4
django-allauth==0.45.0
django-rest-auth==0.9.5
psycopg2==2.9.1